requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.0.6
django-simple-chatbot==0.0.9
djangorestframework==3.13.1
joblib==1.1.0
nltk==3.7
pytz==2022.1
regex==2022.7.9
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
